Refactor each composition class.
