+++ src/components/configure/customkey/TabKey.tsx
@@ -4,11 +4,7 @@ import {
   IMod,
   IModDirection,
   KeycodeCompositionFactory,
-  LayerModComposition,
-  ModsComposition,
   MOD_LEFT,
-  OneShotModComposition,
-  SwapHandsComposition,
 } from '../../../services/hid/Composition';
 
 import AutocompleteKeys from './AutocompleteKeys';
@@ -18,6 +14,10 @@ import { KeyCategory } from '../../../services/hid/KeyCategoryList';
 import { KeyboardLabelLang } from '../../../services/labellang/KeyLabelLangs';
 import { TabKeyActionsType, TabKeyStateType } from './TabKey.container';
 import { IMacroBuffer, MacroBuffer } from '../../../services/macro/Macro';
+import { ModsComposition } from '../../../services/hid/compositions/ModsComposition';
+import { OneShotModComposition } from '../../../services/hid/compositions/OneShotModComposition';
+import { SwapHandsComposition } from '../../../services/hid/compositions/SwapHandsComposition';
+import { LayerModComposition } from '../../../services/hid/compositions/LayerModComposition';
 
 type OwnProps = {
   autoFocus: boolean;
