+++ src/services/hid/Composition.test.ts
@@ -1,40 +1,14 @@
 import {
-  AsciiComposition,
-  BasicComposition,
-  DefLayerComposition,
   IKeycodeCompositionKind,
   KeycodeCompositionFactory,
   KeycodeCompositionKind,
-  LayerModComposition,
-  LayerTapComposition,
-  LayerTapToggleComposition,
-  LooseKeycodeComposition,
-  MacroComposition,
   MOD_ALT,
   MOD_CTL,
   MOD_GUI,
   MOD_LEFT,
   MOD_RIGHT,
   MOD_SFT,
-  ModDirection,
-  ModsComposition,
-  ModTapComposition,
-  MomentaryComposition,
-  OneShotLayerComposition,
-  OneShotModComposition,
-  OP_SH_OFF,
-  OP_SH_OFF_ON,
-  OP_SH_ON,
-  OP_SH_ON_OFF,
-  OP_SH_ONESHOT,
   OP_SH_TAP_TOGGLE,
-  OP_SH_TOGGLE,
-  SwapHandsComposition,
-  TapDanceComposition,
-  ToComposition,
-  ToggleLayerComposition,
-  UnicodeComposition,
-  ViaUserKeyComposition,
 } from './Composition';
 
 const EXPECT_BASIC_LIST = [0b0000_0000_0000_0000, 0b0000_0000_1111_1111];
@@ -66,870 +40,6 @@ const EXPECT_UNKNOWN_LIST = [
 ];
 
 describe('Composition', () => {
-  describe('BasicComposition', () => {
-    test('getCode', () => {
-      let subject = new BasicComposition({
-        code: 0b0000_0100,
-        isAny: false,
-        kinds: ['basic'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0b0000_0100,
-          name: {
-            long: 'KC_A',
-            short: 'KC_A',
-          },
-          label: 'A',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0000_0000_0000_0100);
-      subject = new BasicComposition({
-        code: 0b0000_0000,
-        isAny: false,
-        kinds: ['basic'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0000_0000_0000_0000);
-      subject = new BasicComposition({
-        code: 0b1111_1111,
-        isAny: false,
-        kinds: ['basic'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0000_0000_1111_1111);
-      subject = new BasicComposition({
-        code: 0b1_0000_0000,
-        isAny: false,
-        kinds: ['basic'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0000_0000_0000_0000);
-    });
-  });
-
-  describe('ModsComposition', () => {
-    test('getCode', () => {
-      let subject = new ModsComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0100,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0b0000_0100,
-            name: {
-              long: 'KC_A',
-              short: 'KC_A',
-            },
-            label: 'A',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0001_1111_0000_0100);
-      subject = new ModsComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0000,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0000_1111_0000_0000);
-      subject = new ModsComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0000,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0001_1111_0000_0000);
-      subject = new ModsComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1111_1111,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0000_1111_1111_1111);
-      subject = new ModsComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1111_1111,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0001_1111_1111_1111);
-      subject = new ModsComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1_0000_0000,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0000_1111_0000_0000);
-      subject = new ModsComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1_0000_0000,
-          isAny: false,
-          kinds: ['mods'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0001_1111_0000_0000);
-    });
-  });
-
-  describe('MacroComposition', () => {
-    test('getCode', () => {
-      let subject = new MacroComposition(0b0000_0000);
-      expect(subject.getCode()).toEqual(0b0111_0111_0000_0000);
-      subject = new MacroComposition(0b0111_1111);
-      expect(subject.getCode()).toEqual(0b0111_0111_0111_1111);
-      subject = new MacroComposition(0b1111_1111);
-      expect(subject.getCode()).toEqual(0b0111_0111_0111_1111);
-    });
-  });
-
-  describe('LayerTapComposition', () => {
-    test('getCode', () => {
-      let subject = new LayerTapComposition(2, {
-        code: 0b0000_0100,
-        isAny: false,
-        kinds: ['layer_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0b0000_0100,
-          name: {
-            long: 'KC_A',
-            short: 'KC_A',
-          },
-          label: 'A',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0100_0010_0000_0100);
-      subject = new LayerTapComposition(0, {
-        code: 0b0000_0000,
-        isAny: false,
-        kinds: ['layer_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0100_0000_0000_0000);
-      subject = new LayerTapComposition(15, {
-        code: 0b1111_1111,
-        isAny: false,
-        kinds: ['layer_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0100_1111_1111_1111);
-      subject = new LayerTapComposition(16, {
-        code: 0b1_0000_0000,
-        isAny: false,
-        kinds: ['layer_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0100_0000_0000_0000);
-    });
-  });
-
-  describe('ToComposition', () => {
-    test('getCode', () => {
-      let subject = new ToComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_0000_0100);
-      subject = new ToComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0000_0000);
-      subject = new ToComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_0000_1111);
-      subject = new ToComposition(0b1_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0001_0000);
-      subject = new ToComposition(0b1111_1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_0001_1111);
-    });
-  });
-
-  describe('MomentaryComposition', () => {
-    test('getCode', () => {
-      let subject = new MomentaryComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_0010_0100);
-      subject = new MomentaryComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0010_0000);
-      subject = new MomentaryComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_0010_1111);
-      subject = new MomentaryComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0010_0000);
-    });
-  });
-
-  describe('DefLayerComposition', () => {
-    test('getCode', () => {
-      let subject = new DefLayerComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_0100_0100);
-      subject = new DefLayerComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0100_0000);
-      subject = new DefLayerComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_0100_1111);
-      subject = new DefLayerComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0100_0000);
-    });
-  });
-
-  describe('ToggleLayerComposition', () => {
-    test('getCode', () => {
-      let subject = new ToggleLayerComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_0110_0100);
-      subject = new ToggleLayerComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0110_0000);
-      subject = new ToggleLayerComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_0110_1111);
-      subject = new ToggleLayerComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_0110_0000);
-    });
-  });
-
-  describe('OneShotLayerComposition', () => {
-    test('getCode', () => {
-      let subject = new OneShotLayerComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_1000_0100);
-      subject = new OneShotLayerComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_1000_0000);
-      subject = new OneShotLayerComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_1000_1111);
-      subject = new OneShotLayerComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_1000_0000);
-    });
-  });
-
-  describe('OneShotModComposition', () => {
-    test('getCode', () => {
-      let subject = new OneShotModComposition(ModDirection.right, [
-        MOD_CTL,
-        MOD_SFT,
-        MOD_ALT,
-        MOD_GUI,
-      ]);
-      expect(subject.getCode()).toEqual(0b0101_0010_1011_1111);
-      subject = new OneShotModComposition(ModDirection.left, [
-        MOD_CTL,
-        MOD_SFT,
-        MOD_ALT,
-        MOD_GUI,
-      ]);
-      expect(subject.getCode()).toEqual(0b0101_0010_1010_1111);
-      subject = new OneShotModComposition(ModDirection.right, []);
-      expect(subject.getCode()).toEqual(0b0101_0010_1011_0000);
-      subject = new OneShotModComposition(ModDirection.left, []);
-      expect(subject.getCode()).toEqual(0b0101_0010_1010_0000);
-    });
-  });
-
-  describe('TapDanceComposition', () => {
-    test('getCode', () => {
-      let subject = new TapDanceComposition(2);
-      expect(subject.getCode()).toEqual(0b0101_0111_0000_0010);
-      subject = new TapDanceComposition(0b0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0111_0000_0000);
-      subject = new TapDanceComposition(0b1111_1111);
-      expect(subject.getCode()).toEqual(0b0101_0111_1111_1111);
-      subject = new TapDanceComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0111_0000_0000);
-    });
-  });
-
-  describe('LayerTapToggleComposition', () => {
-    test('getCode', () => {
-      let subject = new LayerTapToggleComposition(0b0100);
-      expect(subject.getCode()).toEqual(0b0101_0010_1100_0100);
-      subject = new LayerTapToggleComposition(0b0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_1100_0000);
-      subject = new LayerTapToggleComposition(0b1111);
-      expect(subject.getCode()).toEqual(0b0101_0010_1100_1111);
-      subject = new LayerTapToggleComposition(0b1_0000_0000);
-      expect(subject.getCode()).toEqual(0b0101_0010_1100_0000);
-    });
-  });
-
-  describe('LayerModComposition', () => {
-    test('getCode', () => {
-      let subject = new LayerModComposition(0b0100, [MOD_CTL, MOD_ALT]);
-      expect(subject.getCode()).toEqual(0b0101_0000_1000_0101);
-      subject = new LayerModComposition(0b0000, []);
-      expect(subject.getCode()).toEqual(0b0101_0000_0000_0000);
-      subject = new LayerModComposition(0b1111, [
-        MOD_CTL,
-        MOD_ALT,
-        MOD_GUI,
-        MOD_SFT,
-      ]);
-      expect(subject.getCode()).toEqual(0b0101_0001_1110_1111);
-      subject = new LayerModComposition(0b1_0000_0000, []);
-      expect(subject.getCode()).toEqual(0b0101_0000_0000_0000);
-    });
-  });
-
-  describe('SwapHandsComposition', () => {
-    test('getCode', () => {
-      let subject = new SwapHandsComposition({
-        code: 0b0000_0100,
-        isAny: false,
-        kinds: ['swap_hands'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0b0000_0100,
-          name: {
-            long: 'KC_A',
-            short: 'KC_A',
-          },
-          label: 'A',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0101_0110_0000_0100);
-      subject = new SwapHandsComposition({
-        code: 0b0000_0000,
-        isAny: false,
-        kinds: ['swap_hands'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0101_0110_0000_0000);
-      subject = new SwapHandsComposition(OP_SH_TOGGLE);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0000);
-      subject = new SwapHandsComposition(OP_SH_TAP_TOGGLE);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0001);
-      subject = new SwapHandsComposition(OP_SH_ON_OFF);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0010);
-      subject = new SwapHandsComposition(OP_SH_OFF_ON);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0011);
-      subject = new SwapHandsComposition(OP_SH_OFF);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0100);
-      subject = new SwapHandsComposition(OP_SH_ON);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0101);
-      subject = new SwapHandsComposition(OP_SH_ONESHOT);
-      expect(subject.getCode()).toEqual(0b0101_0110_1111_0110);
-    });
-  });
-
-  describe('ModTapComposition', () => {
-    test('getCode', () => {
-      let subject = new ModTapComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0100,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0b0000_0100,
-            name: {
-              long: 'KC_A',
-              short: 'KC_A',
-            },
-            label: 'A',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0011_1111_0000_0100);
-      subject = new ModTapComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0000,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0010_1111_0000_0000);
-      subject = new ModTapComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b0000_0000,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0011_1111_0000_0000);
-      subject = new ModTapComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1111_1111,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0010_1111_1111_1111);
-      subject = new ModTapComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1111_1111,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0011_1111_1111_1111);
-      subject = new ModTapComposition(
-        ModDirection.left,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1_0000_0000,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0010_1111_0000_0000);
-      subject = new ModTapComposition(
-        ModDirection.right,
-        [MOD_CTL, MOD_SFT, MOD_ALT, MOD_GUI],
-        {
-          code: 0b1_0000_0000,
-          isAny: false,
-          kinds: ['mod_tap'],
-          direction: MOD_LEFT,
-          modifiers: [],
-          keycodeInfo: {
-            code: 0,
-            name: {
-              long: '',
-              short: '',
-            },
-            label: '',
-            keywords: [],
-          },
-        }
-      );
-      expect(subject.getCode()).toEqual(0b0011_1111_0000_0000);
-      subject = new ModTapComposition(ModDirection.left, [], {
-        code: 0b1111_1111,
-        isAny: false,
-        kinds: ['mod_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0010_0000_1111_1111);
-      subject = new ModTapComposition(ModDirection.right, [], {
-        code: 0b1111_1111,
-        isAny: false,
-        kinds: ['mod_tap'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0011_0000_1111_1111);
-    });
-  });
-
-  describe('UnicodeComposition', () => {
-    test('getCode', () => {
-      let subject = new UnicodeComposition(0b000_0000_0000);
-      expect(subject.getCode()).toEqual(0b1000_0000_0000_0000);
-      subject = new UnicodeComposition(0b111_1111_1111_1111);
-      expect(subject.getCode()).toEqual(0b1111_1111_1111_1111);
-      subject = new UnicodeComposition(0b1000_0000_0000_0000);
-      expect(subject.getCode()).toEqual(0b1000_0000_0000_0000);
-      subject = new UnicodeComposition(0b1111_1111_1111_1111);
-      expect(subject.getCode()).toEqual(0b1111_1111_1111_1111);
-    });
-  });
-
-  describe('LooseKeycodeComposition', () => {
-    test('getCode', () => {
-      let subject = new LooseKeycodeComposition({
-        code: 0b0101_1100_0000_0000,
-        isAny: false,
-        kinds: ['loose_keycode'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0b0101_1100_0000_0000,
-          name: {
-            long: 'RESET',
-            short: 'RESET',
-          },
-          label: 'RESET',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0101_1100_0000_0000);
-      subject = new LooseKeycodeComposition({
-        code: 0b0101_1111_1111_1111,
-        isAny: false,
-        kinds: ['loose_keycode'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0101_1111_1111_1111);
-      subject = new LooseKeycodeComposition({
-        code: 0b0101_1100_1111_0111,
-        isAny: false,
-        kinds: ['loose_keycode'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0101_1100_1111_0111);
-    });
-  });
-
-  describe('AsciiComposition', () => {
-    test('getCode', () => {
-      let subject = new AsciiComposition({
-        code: 33,
-        isAny: false,
-        kinds: ['ascii'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 33,
-          name: {
-            long: '!',
-            short: '!',
-          },
-          label: '!',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(33);
-    });
-  });
-
-  describe('ViaUserKeyComposition', () => {
-    test('getCode', () => {
-      let subject = new ViaUserKeyComposition({
-        code: 0b0111_1110_0000_0000,
-        isAny: false,
-        kinds: ['via_user_key'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0b0111_1110_0000_0000,
-          name: {
-            long: 'Custom User 0',
-            short: 'USER 0',
-          },
-          label: 'User 0',
-          keywords: ['user0'],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0111_1110_0000_0000);
-      subject = new ViaUserKeyComposition({
-        code: 0b0111_1110_0001_1111,
-        isAny: false,
-        kinds: ['via_user_key'],
-        direction: MOD_LEFT,
-        modifiers: [],
-        keycodeInfo: {
-          code: 0,
-          name: {
-            long: '',
-            short: '',
-          },
-          label: '',
-          keywords: [],
-        },
-      });
-      expect(subject.getCode()).toEqual(0b0111_1110_0001_1111);
-    });
-
-    describe('findKeymap', () => {
-      test('customKeycodes is not specified', () => {
-        const actual = ViaUserKeyComposition.findKeymap(
-          0b0111_1110_0000_0000,
-          undefined
-        );
-        expect(actual).not.toBeUndefined();
-        expect(actual!.code).toEqual(0b0111_1110_0000_0000);
-        expect(actual!.desc).toEqual('Kb 0');
-        expect(actual!.kinds).toEqual(['via_user_key']);
-        expect(actual!.keycodeInfo.name.short).toEqual('QK_KB_0');
-        expect(actual!.keycodeInfo.name.long).toEqual('QK_KB_0');
-        expect(actual!.keycodeInfo.label).toEqual('Kb 0');
-      });
-
-      test('customKeycodes is specified', () => {
-        const actual = ViaUserKeyComposition.findKeymap(0b0111_1110_0000_0000, [
-          {
-            name: 'customKeycodeName1',
-            shortName: 'customKeycodeShortName1',
-            title: 'customKeycodeTitle1',
-          },
-        ]);
-        expect(actual).not.toBeUndefined();
-        expect(actual!.code).toEqual(0b0111_1110_0000_0000);
-        expect(actual!.desc).toEqual('customKeycodeTitle1');
-        expect(actual!.kinds).toEqual(['via_user_key']);
-        expect(actual!.keycodeInfo.name.short).toEqual(
-          'customKeycodeShortName1'
-        );
-        expect(actual!.keycodeInfo.name.long).toEqual('customKeycodeName1');
-        expect(actual!.keycodeInfo.label).toEqual('customKeycodeName1');
-      });
-
-      test('customKeycodes is specified but empty', () => {
-        const actual = ViaUserKeyComposition.findKeymap(0b0111_1110_0000_0000, [
-          {
-            name: '',
-            shortName: '',
-            title: '',
-          },
-        ]);
-        expect(actual).not.toBeUndefined();
-        expect(actual!.code).toEqual(0b0111_1110_0000_0000);
-        expect(actual!.desc).toEqual('Kb 0');
-        expect(actual!.kinds).toEqual(['via_user_key']);
-        expect(actual!.keycodeInfo.name.short).toEqual('QK_KB_0');
-        expect(actual!.keycodeInfo.name.long).toEqual('QK_KB_0');
-        expect(actual!.keycodeInfo.label).toEqual('Kb 0');
-      });
-    });
-  });
-
   describe('KeycodeCompositionFactory', () => {
     describe('createBasicComposition', () => {
       test('valid', () => {
